+++ src/main/php/lang/ast/emit/PHP.class.php
@@ -233,13 +233,8 @@ abstract class PHP extends Emitter {
     foreach ($static->initializations as $variable => $initial) {
       $result->out->write('static $'.$variable);
       if ($initial) {
-        if ($this->isConstant($result, $initial)) {
-          $result->out->write('=');
-          $this->emitOne($result, $initial);
-        } else {
-          $result->out->write('= null; null === $'.$variable.' && $'.$variable.'= ');
-          $this->emitOne($result, $initial);
-        }
+        $result->out->write('=');
+        $this->emitOne($result, $initial);
       }
       $result->out->write(';');
     }