+++ src/main/php/lang/ast/emit/PHP81.class.php
@@ -19,7 +19,13 @@ use lang\ast\types\{
  * @see  https://wiki.php.net/rfc#php_81
  */
 class PHP81 extends PHP {
-  use RewriteBlockLambdaExpressions, RewriteDynamicClassConstants, ReadonlyClasses, OmitConstantTypes;
+  use
+    RewriteBlockLambdaExpressions,
+    RewriteDynamicClassConstants,
+    RewriteStaticVariableInitializations,
+    ReadonlyClasses,
+    OmitConstantTypes
+  ;
 
   /** Sets up type => literal mappings */
   public function __construct() {