+++ src/main/php/lang/ast/emit/PHP82.class.php
@@ -19,7 +19,13 @@ use lang\ast\types\{
  * @see  https://wiki.php.net/rfc#php_82
  */
 class PHP82 extends PHP {
-  use RewriteBlockLambdaExpressions, RewriteDynamicClassConstants, ReadonlyClasses, OmitConstantTypes;
+  use
+    RewriteBlockLambdaExpressions, 
+    RewriteDynamicClassConstants, 
+    RewriteStaticVariableInitializations, 
+    ReadonlyClasses, 
+    OmitConstantTypes
+  ;
 
   /** Sets up type => literal mappings */
   public function __construct() {